Rediseno de app

+++ app/src/main/java/com/example/besttravel/models/hotels/HotelsResponse.kt
@@ -8,7 +8,7 @@ data class HotelsResponse (
     @SerializedName("id"          ) var id          : Int?              = null,
     @SerializedName("name"        ) var name        : String?           = null,
     @SerializedName("description" ) var description : String?           = null,
-    @SerializedName("ranking"     ) var ranking     : Int?              = null,
+    @SerializedName("ranking"     ) var ranking     : Double?              = null,
     @SerializedName("address"     ) var address     : String?           = null,
     @SerializedName("phone"       ) var phone       : String?           = null,
     @SerializedName("images"      ) var images      : ArrayList<Images> = arrayListOf(),
